Move Coords Widgets from widget/common to tools/coords
