+++ lib/widgets/tools/coords/base/gcw_coords_output.dart
@@ -1,12 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:gc_wizard/i18n/app_localizations.dart';
-import 'package:gc_wizard/theme/theme.dart';
 import 'package:gc_wizard/widgets/common/base/gcw_button.dart';
-import 'package:gc_wizard/widgets/common/coords/gcw_map_geometries.dart';
 import 'package:gc_wizard/widgets/common/gcw_default_output.dart';
 import 'package:gc_wizard/widgets/common/gcw_tool.dart';
-import 'package:gc_wizard/widgets/common/coords/gcw_mapview.dart';
-import 'package:latlong/latlong.dart';
+import 'package:gc_wizard/widgets/tools/coords/base/gcw_map_geometries.dart';
+import 'package:gc_wizard/widgets/tools/coords/base/gcw_mapview.dart';
 
 class GCWCoordsOutput extends StatefulWidget {
   final String text;
