+++ lib/widgets/tools/coords/base/gcw_mapview.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/plugin_api.dart';
-import 'package:gc_wizard/widgets/common/coords/gcw_map_geometries.dart';
+import 'package:gc_wizard/widgets/tools/coords/base/gcw_map_geometries.dart';
 import 'package:latlong/latlong.dart';
 
 class GCWMapView extends StatefulWidget {
